+++ packages/n8n-node/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the `@probo/n8n-nodes-probo` package will be documented i
 
 ## Unreleased
 
+### Added
+
+- Access Review Source `Get Many` operation to list organization access review sources
+
 ## [0.213.0] - 2026-08-05
 
 ### Added